requirements.txt
asgiref==3.3.1
Django==3.1.4
django-auth-ldap==2.2.0
mysqlclient==2.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap @ file:///D:/Programmation/Web/Tutor/python_ldap-3.3.1-cp38-cp38-win_amd64.whl
pytz==2020.4
sqlparse==0.4.1
